% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-SummarizedExperiment.R
\name{as.SummarizedExperiment}
\alias{as.SummarizedExperiment}
\title{Convert an object to SummarizedExperiment}
\usage{
as.SummarizedExperiment(x, ...)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
\description{
S3 generic for converting objects to
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.
}
