% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{fast.jaccard.r}
\alias{fast.jaccard.r}
\title{Shared nearest neighbors via fast Jaccard index calculation}
\usage{
fast.jaccard.r(.adj.matrix, .prune = 1/15)
}
\arguments{
\item{.adj.matrix}{A non-symetric sparse matrix representation of the nearest neighbors.}

\item{.prune}{A numeric value that determines the threshold for pruning small values from the Jaccard index matrix.}
}
\value{
A shared nearest neighbors sparse adjacency matrix of Jaccard indices.
}
\description{
This function calculates the Jaccard index between cells based on their nearest neighbors. The function is optimized for speed and memory efficiency.
}
