% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{get.adj.matrix}
\alias{get.adj.matrix}
\title{Sparse matrix representation of nearest neighbors}
\usage{
get.adj.matrix(.nn.idx)
}
\arguments{
\item{.nn.idx}{A matrix where each row represents a cell and each column represents a nearest neighbor. The elements of the matrix are the indices of the nearest neighbors.}
}
\value{
A non-symetric sparse matrix representation of the nearest neighbors.
}
\description{
This function converts a matrix of nearest neighbors into a non-symetric sparse matrix representation.
}
