% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get.cells.list.mat}
\alias{get.cells.list.mat}
\title{Create .cells object from count matrices}
\usage{
get.cells.list.mat(.count.mat.list, .meta, .compress = FALSE, .verbose = TRUE)
}
\arguments{
\item{.count.mat.list}{A named list containing count matrices for each sample. List names correspond to sample names and must match rownames of .meta. Each element can be a matrix, dgCMatrix, or similar matrix-like object.}

\item{.meta}{A data frame containing the sample-wise metadata. Rownames must match names of .count.mat.list.}

\item{.compress}{A logical specifying whether to compress the RDS files. Default is FALSE for faster I/O.}

\item{.verbose}{A logical specifying whether to print progress messages. Default is TRUE.}
}
\value{
A named list containing the on-disk location of RDS files containing one expression matrix for each sample, suitable for use with setup.lm.obj.
}
\description{
This function facilitates the creation of a .cells object suitable for setup.lm.obj from count matrices already loaded into memory.
}
\examples{
\dontrun{
# Fetch example data
trajectory_data <- fetch_trajectory_data()
sim_trajectory.meta <- trajectory_data$meta
sim_trajectory <- trajectory_data$SCE
  
# Process metadata for tinydenseR format
sim_trajectory.meta <- sim_trajectory.meta[, c("Condition", "Replicate", "Sample")] |>
  unique()
rownames(sim_trajectory.meta) <- sim_trajectory.meta$Sample

# Create .cells object using SCE method for subset
.min.meta <- sim_trajectory.meta[c("A_R1", "B_R1"), ]
.min.cells <- get.cells.SCE(.sce.obj = sim_trajectory,
                            .sample.var = "Sample") |>
  (\(x) x[.min.meta$Sample])()

# Create .cells object
cells <- get.cells.list.mat(.count.mat.list = count.matrices,
                   .meta = sim_trajectory.meta[c("A_R1", "B_R1"), ])

# Use with setup.lm.obj
lm.obj <- setup.lm.obj(.cells = cells,
                       .meta = sim_trajectory.meta[c("A_R1", "B_R1"), ],
                       .assay.type = "RNA")
}
}
