% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\name{get.landmarks}
\alias{get.landmarks}
\title{Get landmarks}
\usage{
get.landmarks(
  .lm.obj,
  .verbose = TRUE,
  .seed = 123,
  .nHVG = 5000,
  .nPC = 30,
  .exc.vdj.mito.ribo.genes.from.hvg = TRUE,
  .force.in = NULL
)
}
\arguments{
\item{.lm.obj}{An object initialized with setup.lm.obj.}

\item{.verbose}{A logical specifying whether to print progress messages. Default is TRUE.}

\item{.seed}{An integer specifying the .seed for reproducibility. Default is 123.}

\item{.nHVG}{An integer specifying the number of highly variable genes to use for landmark identification. Default is 5000}

\item{.nPC}{An integer specifying the number of principal components to use for landmark identification. Default is 30.}

\item{.exc.vdj.mito.ribo.genes.from.hvg}{A logical specifying whether to exclude VDJ, mitochondrial, and ribosomal genes from the highly variable genes. Default is TRUE.}

\item{.force.in}{A character vector specifying the genes to force into the landmark set. Default is NULL. Applies only if \code{assay.type} is \code{"RNA"} since all features are included for \code{"cyto"}.}
}
\value{
The .lm.obj now containing the landmarks, a key linking each landmark to a row of .metadata and the PCA object.
}
\description{
This function identifies landmarks from the input expression matrices.
}
