% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.model.R
\name{get.stats}
\alias{get.stats}
\title{Get stats}
\usage{
get.stats(
  .lm.obj,
  .design,
  .contrasts = NULL,
  .block = NULL,
  .verbose = TRUE,
  .seed = 123
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks, get.graph and get.map.}

\item{.design}{A matrix specifying the design matrix to be used.}

\item{.contrasts}{A matrix specifying the contrasts to be used.}

\item{.block}{A character vector of length one indicating which column of \code{.lm.obj$metadata} to be used as blocking factor.}

\item{.verbose}{A logical indicating whether to print progress messages.}

\item{.seed}{An integer to set the seed for reproducibility.}
}
\value{
The .lm.obj now containing the stats from fitting a voomLmFit model.
}
\description{
Fits a model of outcomes as a function of the sample wise sum of cell-landmark neighboring probabilities
}
