% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTDR.R
\name{GetTDR}
\alias{GetTDR}
\alias{GetTDR.default}
\alias{GetTDR.Seurat}
\alias{GetTDR.SingleCellExperiment}
\title{Extract a TDRObj from a container object}
\usage{
GetTDR(x, ...)

\method{GetTDR}{default}(x, ...)

\method{GetTDR}{Seurat}(x, ...)

\method{GetTDR}{SingleCellExperiment}(x, ...)
}
\arguments{
\item{x}{An object that may contain a TDRObj.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{\linkS4class{TDRObj}}.
}
\description{
S3 generic that retrieves a \code{\linkS4class{TDRObj}} from the
object in which \code{\link{RunTDR}} stored it.
}
\section{Methods (by class)}{
\itemize{
\item \code{GetTDR(default)}: Default method – returns a TDRObj as-is, errors otherwise

\item \code{GetTDR(Seurat)}: Extract TDRObj from a Seurat object's Misc slot

\item \code{GetTDR(SingleCellExperiment)}: Extract TDRObj from a SingleCellExperiment's metadata

}}
