% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goi.R
\name{goi.summary}
\alias{goi.summary}
\title{Creates several summaries from a list of genes of interest.}
\usage{
goi.summary(
  .lm.obj,
  .goi,
  .id.idx = NULL,
  .id = NULL,
  .id.from = "clustering",
  .verbose = TRUE
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with \code{get.graph} and \code{get.map}.}

\item{.goi}{A character vector of genes of interest.}

\item{.id.idx}{An index of the cluster or celltype. If NULL, all landmarks are used.}

\item{.id}{The id of the cluster or celltype. If NULL, all landmarks are used.}

\item{.id.from}{The source of the id. Can be set to "clustering" or "celltyping". Defaults to NULL.}

\item{.verbose}{Logical indicating whether to print verbose output.}
}
\value{
The .lm.obj with the goi field populated with the updated cluster/celltype names.
}
\description{
This function creates several summaries typically used in scRNAseq data
analysis from  a list of genes of interest (GOI).
}
