% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDRObj-class.R
\name{is.TDRObj}
\alias{is.TDRObj}
\title{Check if an object is a TDRObj}
\usage{
is.TDRObj(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical: TRUE if \code{x} is a TDRObj, FALSE otherwise.
}
\description{
Check if an object is a TDRObj
}
