% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{leiden.cluster}
\alias{leiden.cluster}
\title{Leiden clustering}
\usage{
leiden.cluster(
  .lm.obj = NULL,
  .sim.matrix,
  .resolution.parameter,
  .small.size = floor(nrow(.sim.matrix)/200),
  .verbose = TRUE,
  .seed = 123
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks. If provided, the function will use the Laplacian Eigenmap embedding from the .lm.obj for initialization. If NULL, the function will compute a 3-dimensional embedding using irlba on the similarity matrix.}

\item{.sim.matrix}{A square matrix representing the graph of the cells. Elements of the matrix should be a similarity measure of pairs of observations.}

\item{.resolution.parameter}{A numeric value that determines the granularity of the clustering. Higher values result in more clusters.}

\item{.small.size}{An integer value that determines the size of clusters that are considered "stragglers". Stragglers are assigned to the cluster they are most connected to. Defaults to 0.5\% of the number of cells.}

\item{.verbose}{A logical value indicating whether to print messages about the clustering process.}

\item{.seed}{An integer value that sets the seed for reproducibility.}
}
\value{
A factor vector of cluster assignments.
}
\description{
This function performs Leiden clustering on a graph of cells. The graph is represented by an adjacency matrix, where the elements of the matrix are the likelihood that pairs of vertices are adjacent. The function returns a factor vector of cluster assignments.
}
