% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{list_celltyping_solutions}
\alias{list_celltyping_solutions}
\alias{list_celltyping_solutions.TDRObj}
\title{List stored celltyping solutions}
\usage{
list_celltyping_solutions(x, ...)

\method{list_celltyping_solutions}{TDRObj}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{TDRObj}}, Seurat, or SingleCellExperiment
object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A character vector of solution names.
}
\description{
Returns the names of all stored celltyping solutions (excluding the
active \code{$ids} slot).
}
\examples{
\dontrun{
list_celltyping_solutions(tdr.obj)
# [1] "cell_type_l1" "cell_type_l2" "azimuth_pred"
}

}
\seealso{
\code{\link{set_active_celltyping}}, \code{\link{import_cell_annotations}}
}
