% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R, R/lm.graph.embed.R
\name{set_active_clustering.Seurat}
\alias{set_active_clustering.Seurat}
\alias{set_active_clustering.SingleCellExperiment}
\alias{set_active_clustering}
\alias{set_active_clustering.TDRObj}
\title{Set active clustering solution}
\usage{
\method{set_active_clustering}{Seurat}(x, ...)

\method{set_active_clustering}{SingleCellExperiment}(x, ...)

set_active_clustering(x, ...)

\method{set_active_clustering}{TDRObj}(x, .column.name, .verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{TDRObj}}, Seurat, SingleCellExperiment, or HDF5AnnData
(anndataR) object.}

\item{...}{Additional arguments passed to methods.}

\item{.column.name}{Character: name of the stored clustering solution to activate.}

\item{.verbose}{Logical (default TRUE).}
}
\value{
Updated object with the selected clustering as active \code{$ids}.
}
\description{
Switches the active clustering to a previously stored solution and
refreshes all clustering-dependent downstream slots.
}
