% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\name{SetTDR}
\alias{SetTDR}
\alias{SetTDR.default}
\alias{SetTDR.Seurat}
\alias{SetTDR.SingleCellExperiment}
\alias{SetTDR.SummarizedExperiment}
\title{Store a TDRObj inside a container object}
\usage{
SetTDR(x, tdr, ...)

\method{SetTDR}{default}(x, tdr, ...)

\method{SetTDR}{Seurat}(x, tdr, ...)

\method{SetTDR}{SingleCellExperiment}(x, tdr, ...)

\method{SetTDR}{SummarizedExperiment}(x, tdr, ...)
}
\arguments{
\item{x}{The container (Seurat, SCE, or TDRObj).}

\item{tdr}{A \code{\linkS4class{TDRObj}} to store.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The container with the TDRObj stored.
}
\description{
Store a TDRObj inside a container object
}
\section{Methods (by class)}{
\itemize{
\item \code{SetTDR(default)}: Default: if x is a TDRObj, returns tdr directly

\item \code{SetTDR(Seurat)}: Store TDRObj in Seurat Misc slot

\item \code{SetTDR(SingleCellExperiment)}: Store TDRObj in SCE metadata

\item \code{SetTDR(SummarizedExperiment)}: Store TDRObj in SummarizedExperiment metadata

}}
