% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinydenseR-package.R
\docType{package}
\name{tinydenseR-package}
\alias{tinydenseR}
\alias{tinydenseR-package}
\title{tinydenseR: Linking Cell-To-Cell Variation to Sample-to-Sample Variation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

tinydenseR is a landmark-based platform for single cell data analysis that identifies differential cell states and features, including subtle within-cluster state changes. Modeling samples as replicates, tinydenseR enhances analytic efficiency and reproducibility while preserving the richness of single cell data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://opensource.nibr.com/tinydenseR/}
  \item Report bugs at \url{https://github.com/Novartis/tinydenseR/issues}
}

}
\author{
\strong{Maintainer}: Pedro Milanez-Almeida \email{pedro.milanez@novartis.com} (\href{https://orcid.org/0000-0001-8285-957X}{ORCID})

Authors:
\itemize{
  \item Pedro Milanez-Almeida \email{pedro.milanez@novartis.com} (\href{https://orcid.org/0000-0001-8285-957X}{ORCID})
}

}
\keyword{internal}
