% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{celltyping}
\alias{celltyping}
\title{Map clusters to celltypes.}
\usage{
celltyping(.lm.obj, .celltyping.map)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks and get.graph.}

\item{.celltyping.map}{A named list, where names are celltypes and each elemt is a character vector of cluster mapping to each respective celltype. A celltype can be comprised of more than one cluster, but each cluser can only be mapped to one celltype.}
}
\value{
The .lm.obj with the celltyping field populated with the celltype mapping, mean celltype expression values as well as the mean expression heatmap.
}
\description{
This function maps clusters to cell types.
}
