% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{.tdr_cache_read}
\alias{.tdr_cache_read}
\title{Read a cached slot from disk}
\usage{
.tdr_cache_read(meta_record)
}
\arguments{
\item{meta_record}{A metadata list produced by \code{.tdr_cache_write}.}
}
\value{
The deserialized R object.
}
\description{
Read a cached slot from disk
}
\keyword{internal}
