% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get.cells.SCE}
\alias{get.cells.SCE}
\title{Create .cells object from SingleCellExperiment object}
\usage{
get.cells.SCE(
  .sce.obj,
  .meta,
  .sample.var,
  .assay = "counts",
  .min.cells.per.sample = 10,
  .compress = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{.sce.obj}{A SingleCellExperiment object containing expression data.}

\item{.meta}{A data frame containing the sample-wise metadata. Row names must be sample names matching elements of the object metadata specified by \code{.sample.var}. Required for filtering valid samples.}

\item{.sample.var}{A character string specifying the metadata column to use for sample identification. This column will be used to split cells into samples.}

\item{.assay}{A character string specifying the assay to use. Default is "counts".}

\item{.min.cells.per.sample}{An integer specifying the minimum number of cells required per sample. Samples with fewer cells will be excluded. Default is 10.}

\item{.compress}{A logical specifying whether to compress the RDS files. Default is FALSE for faster I/O.}

\item{.verbose}{A logical specifying whether to print progress messages. Default is TRUE.}
}
\value{
A named list containing the on-disk location of RDS files containing one expression matrix for each sample, suitable for use with setup.lm.obj.
}
\description{
This function facilitates the creation of a .cells object suitable for setup.lm.obj from a SingleCellExperiment object.
}
\examples{
\dontrun{
# Fetch example data
trajectory_data <- fetch_trajectory_data()
sim_trajectory.meta <- trajectory_data$meta
sim_trajectory <- trajectory_data$SCE

# Process metadata for tinydenseR format
sim_trajectory.meta <- sim_trajectory.meta[, c("Condition", "Replicate", "Sample")] |>
  unique()
rownames(sim_trajectory.meta) <- sim_trajectory.meta$Sample

# Use the SCE object directly
cells <- get.cells.SCE(.sce.obj = sim_trajectory,
                       .sample.var = "Sample")

# Use with setup.lm.obj
lm.obj <- setup.lm.obj(.cells = cells,
                       .meta = sim_trajectory.meta,
                       .assay.type = "RNA")
}
}
