% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.model.R
\name{get.dea}
\alias{get.dea}
\title{Differential Expression Analysis}
\usage{
get.dea(
  .lm.obj,
  .design,
  .contrasts = NULL,
  .block = NULL,
  .geneset.ls = NULL,
  .id.idx = NULL,
  .id = NULL,
  .id.from = NULL,
  .verbose = TRUE
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with
get.landmarks, get.graph, get.map and get.stats.}

\item{.design}{A matrix specifying the design matrix to be used.}

\item{.contrasts}{A matrix specifying the contrasts to be used.}

\item{.block}{A character vector of length one indicating which column of \code{.lm.obj$metadata} to be used as blocking factor.}

\item{.geneset.ls}{A list of character vectors specifying the gene sets to be used in differential gene set score analysis (via pseudobulk GSVA). Default is NULL (no gene set scoring is run).}

\item{.id.idx}{A list of integer vectors specifying the indices of the landmarks to be used. Default is NULL. If provided, cells in each sample that have the selected landmark as first neighbor are included.}

\item{.id}{A character vector specifying the ids to be used. Default is NULL.}

\item{.id.from}{A character specifying the source of the ids. Default is NULL.
Options are "clustering" and "celltyping".}

\item{.verbose}{A logical indicating whether to print progress messages. Default is TRUE.}
}
\value{
The .lm.obj now containing the stats from fitting a linear model for each marker in each cluster.
}
\description{
This function performs differential expression analysis for each feature
using .coef as response. If provided, only cells within .id are used.
}
