% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{get.graph}
\alias{get.graph}
\title{Graph embedding of landmarks}
\usage{
get.graph(
  .lm.obj,
  .k = 20,
  .scale = if (.lm.obj$assay.type == "RNA") FALSE else TRUE,
  .verbose = TRUE,
  .seed = 123,
  .cl.method = "snn",
  .cl.resolution.parameter = 0.8,
  .small.size = floor(x = nrow(x = .lm.obj$lm)/200)
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks.}

\item{.k}{An integer value that determines the number of nearest neighbors to consider in the graph.}

\item{.scale}{A logical value indicating whether to scale the data before embedding. Default is FALSE for assay type RNA (to avoid scaling PCA results) and TRUE for cyto.}

\item{.verbose}{A logical value indicating whether to print messages about the embedding process.}

\item{.seed}{An integer value that sets the seed for reproducibility.}

\item{.cl.method}{A character value that determines the clustering method. Options are "snn" for shared nearest neighbors and "fgraph" for fuzzy graph.}

\item{.cl.resolution.parameter}{A numeric value that determines the granularity of the clustering. Higher values result in more clusters.}

\item{.small.size}{An integer value that determines the size of clusters that are considered "stragglers". Stragglers are assigned to the cluster they are most connected to. Defaults to 0.5\% of the number of cells.}
}
\value{
The .lm.obj now containing a graph object with, for example, the UMAP embedding, cluster assignments, mean cluster expression values as well as the mean expression heatmap.
}
\description{
This function performs graph-based clustering, dimensionality reduction, and map building of landmark
}
