% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{get.lm.features.stats}
\alias{get.lm.features.stats}
\title{Graph-based feature discovery for landmarks}
\usage{
get.lm.features.stats(.lm.obj)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks.}
}
\value{
A list that includes the feature discovery results for each landmark.
}
\description{
This function identifies features that are most characteristic for each landmark based on the PCA rotation matrix.
}
