% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{get.map}
\alias{get.map}
\title{Mapping cells to landmarks}
\usage{
get.map(
  .lm.obj,
  .ref.obj = NULL,
  .integrate.vars = NULL,
  .celltype.col.name = "cell_type",
  .cl.ct.to.ign = NULL,
  .irrel.par = NULL,
  .verbose = TRUE,
  .seed = 123
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks and get.graph.}

\item{.ref.obj}{A symphony object with Z_corr field populated.}

\item{.integrate.vars}{A character vector of batch variables to integrate (column names in metadata).}

\item{.celltype.col.name}{The column name in the reference object metadata that contains cell type information.}

\item{.cl.ct.to.ign}{A character value that determines the cluster or cell type to ignore during mapping. Default is NULL. Use this to ignore clusters or cell types that are not relevant for the statistical analysis, such as erythrocytes in PBMC or BMMC, for example, wheresince their presence in such samples is more likely to be associated with technical artifacts during sample processing in the lab rather than biological phenotypes.}

\item{.irrel.par}{A character vector of irrelevant parameters to exclude from the mapping. Default is NULL.}

\item{.verbose}{A logical value indicating whether to print messages about the mapping process.}

\item{.seed}{An integer value that sets the seed for reproducibility.}
}
\value{
The .lm.obj now containing a list that includes the sample-wise sum of cell-landmark neighboring probabilities, celltypes (if available, see "celltyping") and cluster assignments for each cell, and nearest neighboring landmarks for each cell.
}
\description{
This function maps cells to landmarks to derive the probability that each cell is in the neighborhood of a landmark.
}
