% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.model.R
\name{get.marker}
\alias{get.marker}
\title{Differential Expression Analysis Comparing Cell Subsets}
\usage{
get.marker(
  .lm.obj,
  .geneset.ls = NULL,
  .id1.idx = NULL,
  .id2.idx = NULL,
  .id1 = NULL,
  .id2 = "..all.other.subsets..",
  .id.from = "clustering"
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with
get.landmarks, get.graph, get.map and get.stats.}

\item{.geneset.ls}{A list of character vectors specifying the gene sets to be used in differential gene set score analysis (via pseudobulk GSVA). Default is NULL (no gene set scoring is run).}

\item{.id1.idx}{A list of integer vectors specifying the indices of the landmarks to be used for .id1. Default is NULL.}

\item{.id2.idx}{A list of integer vectors specifying the indices of the landmarks to be used for .id2. Default is NULL.}

\item{.id1}{A character vector specifying the ids to be used.}

\item{.id2}{A character vector specifying the ids to be used. Default is \code{..all.other.subsets..}.}

\item{.id.from}{A character specifying the source of the ids. Options are "clustering" and "celltyping".}
}
\value{
The .lm.obj now containing the stats from fitting a linear model for each marker in each cluster.
}
\description{
This function performs differential expression analysis for each feature
comparing cell subsets.
}
