% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get.meta.SCE}
\alias{get.meta.SCE}
\title{Extract metadata from SingleCellExperiment object}
\usage{
get.meta.SCE(.sce.obj, .sample.var, .verbose = TRUE)
}
\arguments{
\item{.sce.obj}{A SingleCellExperiment object containing cell metadata.}

\item{.sample.var}{A character string specifying the metadata column to use for sample identification.}

\item{.verbose}{A logical specifying whether to print progress messages. Default is TRUE.}
}
\value{
A data frame containing sample-wise metadata with sample names as rownames.
}
\description{
This function extracts sample-wise metadata from a SingleCellExperiment object.
}
\examples{
\dontrun{
# Assuming you have a SingleCellExperiment object with sample metadata
meta <- get.meta.SCE(.sce.obj = sce.obj,
                     .sample.var = "sample_id")
}
}
