% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{interactFeatPlot}
\alias{interactFeatPlot}
\title{Plot landmark feature discovery summary}
\usage{
interactFeatPlot(
  .lm.obj,
  .n.feat = 10,
  .plot.dims = "umap",
  .PC.x = 1,
  .PC.y = 2,
  .panel.size = 1.5,
  .point.size = 2
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks, get.graph and get.lm.features.}

\item{.n.feat}{The number of top features to plot. Defaults to 10.}

\item{.plot.dims}{The dimensions to plot. Can be set to "umap", "pca", or a combination of two column names from .lm.obj$lm. Defaults to "umap".}

\item{.PC.x}{The principal component to plot on the x-axis. Defaults to 1 (PC1).}

\item{.PC.y}{The principal component to plot on the y-axis. Defaults to 2 (PC2).}

\item{.panel.size}{The size of the panel.}

\item{.point.size}{The size of the points in the plot.}
}
\value{
A ggiraph object if the ggiraph package is available, otherwise a ggplot object.
}
\description{
This function plots the summary of the landmark feature discovery.
}
\note{
Interactive features require the ggiraph package. If ggiraph is not available, a static plot will be returned with a warning.
}
