% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.graph.embed.R
\name{lm.cluster}
\alias{lm.cluster}
\title{Leiden clustering of landmarks}
\usage{
lm.cluster(
  .lm.obj,
  .cl.method = "snn",
  .cl.resolution.parameter = 0.8,
  .seed = 123,
  .verbose = TRUE,
  .small.size = 3
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj.}

\item{.cl.method}{A character value that determines the clustering method. Options are "snn" for shared nearest neighbors and "fgraph" for fuzzy graph.}

\item{.cl.resolution.parameter}{A numeric value that determines the granularity of the clustering. Higher values result in more clusters.}

\item{.seed}{An integer value that sets the seed for reproducibility.}

\item{.verbose}{A logical value indicating whether to print messages about the clustering process.}

\item{.small.size}{An integer value that determines the size of clusters that are considered "stragglers". Stragglers are assigned to the cluster they are most connected to.}
}
\value{
The .lm.obj now containing a list that includes the cluster assignments for each landmark, mean cluster expression values as well as the mean expression heatmap.
}
\description{
This function performs Leiden clustering on the landmarks.
}
