% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot2Markers}
\alias{plot2Markers}
\title{Bidimensional plots for markers expression visualization}
\usage{
plot2Markers(
  .lm.obj,
  .id = NULL,
  .id.from = "clustering",
  .x.feature = "CD3",
  .y.feature = "CD20",
  .bins = 128,
  .legend.position = "right",
  .plot.title = "",
  .panel.size = 1.5,
  .reference = TRUE,
  .density.bins = 32,
  .sd.range = c(-3, 6)
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks and get.graph.}

\item{.id}{The id of the cluster or celltype to plot. If NULL, all landmarks are plotted.}

\item{.id.from}{The source of the id. Can be set to "clustering" or "celltyping". Defaults to "clustering".}

\item{.x.feature}{The x-axis marker. Defaults to "CD3".}

\item{.y.feature}{The y-axis marker. Defaults to "CD20".}

\item{.bins}{The number of bins for the hexagonal plot. Defaults to 128.}

\item{.legend.position}{The position of the legend. Defaults to "right" and can be set to "left", "top", "bottom" or "none".}

\item{.plot.title}{The title of the plot.}

\item{.panel.size}{The size of the panel.}

\item{.reference}{If TRUE, the plot will show the density of the landmarks. Defaults to TRUE.}

\item{.density.bins}{The number of bins for the density plot. Defaults to 32.}

\item{.sd.range}{The range of the standard deviation outside which cells are considered outliers and excluded from plot. Defaults to c(-3, 6).}
}
\description{
This function plots the expression of two markers in a bidimensional plot.
}
