% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotBeeswarm}
\alias{plotBeeswarm}
\title{Bee swarm Plot of Density Estimate Change}
\usage{
plotBeeswarm(
  .lm.obj,
  .stats.obj,
  .coefs,
  .q = 0.1,
  .q.from = "pca.weighted.q",
  .split.by = if (length(x = .coefs) > 1) "none" else "clustering",
  .swarm.title = NULL,
  .label.substr.rm = "",
  .point.size = 0.1,
  .facet.scales = "fixed",
  .row.space.scaler = 0.2,
  .col.space.scaler = 0.1,
  .panel.width = 1.5,
  .legend.position = "right",
  .perc.plot = TRUE
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks, get.graph and get.map.}

\item{.stats.obj}{The list object that is the result of running get.stats.}

\item{.coefs}{The coefficient level to plot.}

\item{.q}{The q-value threshold to use for coloring the plot. Defaults to 0.1.}

\item{.q.from}{The source of the q values. Can be set to "density.weighted.bh.fdr", or "pca.weighted.q". Defaults to "pca.weighted.q".}

\item{.split.by}{The variable to split the plot by. Can be set to "none", "clustering", or "celltyping". If either of the two latter, .coefs must be provided and, if not, defaults to showing the first coefficient in formula. Defaults to "none".}

\item{.swarm.title}{The title of the swarm plot. If NULL and if there are no facets (facets are only used if length of \code{.coefs} is greater than 1 and pops are split by clustering or celltyping), the title will be set to the value of .coefs.}

\item{.label.substr.rm}{The substring to remove from the labels.}

\item{.point.size}{The size of the points in the plot. Defaults to 0.1.}

\item{.facet.scales}{The scales to use. Defaults to "fixed".}

\item{.row.space.scaler}{The scaler for the row space. Defaults to 0.2.}

\item{.col.space.scaler}{The scaler for the column space. Defaults to 0.1.}

\item{.panel.width}{The width of the panel. Defaults to 1.5 and is used only if the length of \code{.coefs} is greater than 1.}

\item{.legend.position}{The position of the legend. Defaults to "bottom" and can be set to "left", "top", "bottom", "right" or "none".}

\item{.perc.plot}{If TRUE, the percentages of cells in each cluster or celltype are plotted. Defaults to TRUE. If FALSE, only the beeswarm plot is shown.}
}
\description{
This function plots the statistics of modeling outcomes as a function of the sample wise sum of cell-landmark neighboring probabilities
}
