% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDEA}
\alias{plotDEA}
\title{Plot Differential Expression Analysis Results (Deprecated)}
\usage{
plotDEA(
  .tdr.obj,
  .dea.obj,
  .coefs = colnames(x = .dea.obj$coefficients),
  .order.by = "clustering",
  .markers = colnames(x = .tdr.obj@graph[[.order.by]]$median.exprs),
  .q = 0.1,
  .row.space.scaler = 0.2,
  .col.space.scaler = 0.065,
  .label.substr.rm = ""
)
}
\arguments{
\item{.tdr.obj}{A tinydenseR object.}

\item{.dea.obj}{Differential expression results from \code{get.dea()} or \code{get.pbDE()}.}

\item{.coefs}{Character vector of coefficient names to plot.}

\item{.order.by}{Character: "clustering" or "celltyping".}

\item{.markers}{Character vector of feature names.}

\item{.q}{Numeric adjusted p-value threshold.}

\item{.row.space.scaler}{Numeric row height scaling.}

\item{.col.space.scaler}{Numeric column width scaling.}

\item{.label.substr.rm}{Character substring to remove from labels.}
}
\value{
A \code{ggplot} heatmap.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plotDEA()} has been renamed to \code{\link{plotPbDE}()} for clarity. This function
is provided for backward compatibility and will be removed in a future version.
}
\seealso{
\code{\link{plotPbDE}}
}
\keyword{internal}
