% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDEA}
\alias{plotDEA}
\title{Plot Differential Expression Analysis Results}
\usage{
plotDEA(
  .lm.obj,
  .dea.obj,
  .coefs,
  .order.by = "clustering",
  .markers = colnames(x = .lm.obj$graph[[.order.by]]$mean.exprs),
  .q = 0.1,
  .row.space.scaler = 0.2,
  .col.space.scaler = 0.065,
  .label.substr.rm = ""
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with \code{get.landmarks}, \code{get.graph} and \code{get.map}.}

\item{.dea.obj}{The list object that is the result of running \code{get.dea}.}

\item{.coefs}{The coefficient level to plot.}

\item{.order.by}{The variable to order the row (proteins) by. Can be set to \code{"clustering"} or \code{"celltyping"}. Defaults to \code{"clustering"}.}

\item{.markers}{A character vector with the markers to plot. Defaults to the column names of the \code{"clustering"} or \code{"celltyping"} mean expression matrix.}

\item{.q}{The q-value threshold to use for coloring the plot.}

\item{.row.space.scaler}{The scaler for the row space. Defaults to \code{0.2}.}

\item{.col.space.scaler}{The scaler for the column space. Defaults to \code{0.5}.}

\item{.label.substr.rm}{The substring to remove from the labels.}
}
\description{
This function plots the statistics of modeling outcomes as a function of feature (e.g., genes or proteins) expression.
}
