% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(.lm.obj, .id.from = "clustering")
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks and get.graph.}

\item{.id.from}{Either \code{"clustering"} or \code{"celltyping"}. Defaults to \code{"clustering"}.}
}
\value{
Returns a \code{gtable}, \code{gTree}, \code{grob}, \code{gDesc} object created with \code{pheatmap} and draws on the current device.
}
\description{
This function plots a heatmap of mean expression values across clusters or cell types. For cytometry experiments, all markers in \code{.lm.obj$marker} are returned,  while the top 3 positive and negative genes contributing to each principal component are returned for scRNA-seq experiments.
}
