% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSamplePCA}
\alias{plotSamplePCA}
\title{Plot Sample PCA}
\usage{
plotSamplePCA(
  .lm.obj,
  .labels.from = colnames(x = .lm.obj$metadata)[1],
  .cat.feature.color = Color.Palette[1, 1:5],
  .point.size = 1,
  .panel.size = 2,
  .midpoint = if (is.numeric(x = .lm.obj$metadata[[.labels.from]])) stats::median(x =
    .lm.obj$metadata[[.labels.from]]) else NA,
  .seed = 123
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks, get.graph and get.map.}

\item{.labels.from}{The source of the labels. Must be a .lm.obj$metadata column name. Defaults to the first column name of .lm.obj$metadata.}

\item{.cat.feature.color}{The color palette for the categories. Defaults to the Color.Palette.}

\item{.point.size}{The size of the points in the plot.}

\item{.panel.size}{The size of the panel.}

\item{.midpoint}{The midpoint for the color gradient. Defaults to the median of the \code{.labels.from} column.}

\item{.seed}{The .seed for the random number generator.}
}
\description{
This function plots the PCA of samples in abundance estimate space.
}
