% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTradPerc}
\alias{plotTradPerc}
\title{Plot Traditional Percentages}
\usage{
plotTradPerc(
  .lm.obj,
  .x.split = colnames(x = .lm.obj$metadata)[1],
  .pop = NULL,
  .pop.from = "clustering",
  .line.by = NULL,
  .dodge.by = NULL,
  .x.space.scaler = 0.25,
  .height = 1.5,
  .cat.feature.color = Color.Palette[1, 1:5],
  .seed = 123,
  .orientation = "wide"
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks, get.graph and get.map.}

\item{.x.split}{The column of metadata to use in the x axis. Defaults to the first column of .lm.obj$metadata.}

\item{.pop}{The name of the pop to plot in the y axis.}

\item{.pop.from}{Either \code{"clustering"} or \code{"celltyping"}. Defaults to \code{"clustering"}.}

\item{.line.by}{If provided, draws a line connecting dots from the same subject.}

\item{.dodge.by}{The column of metadata to use for coloring the points. Defaults to NULL, which makes all points \code{"black"}.}

\item{.x.space.scaler}{The scaler for the x axis Defaults to \code{0.5}.}

\item{.height}{The height of the plot. Defaults to \code{1.5}.}

\item{.cat.feature.color}{The color palette for the categories. Defaults to the Color.Palette.}

\item{.seed}{The seed for the random number generator that control x-axis jitter. Defaults to \code{123}.}

\item{.orientation}{The orientation of the plot. Can be set to \code{"wide"} or \code{"square"}. Defaults to \code{"wide"}.}
}
\description{
This function plots the actual percentages from traditional analysis to facilitate visual inspection of value distributions.
}
