% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotTradStats}
\alias{plotTradStats}
\title{Plot Traditional Statistics}
\usage{
plotTradStats(
  .lm.obj,
  .stats.obj,
  .split.by = "clustering",
  .coefs = colnames(x = .stats.obj$trad[[.split.by]]$fit$coefficients),
  .q = 0.1,
  .row.space.scaler = 0.2,
  .col.space.scaler = 0.07,
  .label.substr.rm = ""
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks, get.graph and get.map.}

\item{.stats.obj}{The list object that is the result of running get.stats.}

\item{.split.by}{The variable to split the plot by. Can be set to "clustering" or "celltyping". Defaults to "clustering".}

\item{.coefs}{The coefficient level to plot.}

\item{.q}{The q-value threshold to use for coloring the plot.}

\item{.row.space.scaler}{The scaler for the row space. Defaults to 0.2.}

\item{.col.space.scaler}{The scaler for the column space. Defaults to 0.5.}

\item{.label.substr.rm}{The substring to remove from the labels.}
}
\description{
This function plots the statistics of modeling outcomes as a function of the abundance of the cells in each cluster or celltype.
}
