% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP}
\usage{
plotUMAP(
  .lm.obj,
  .feature = .lm.obj$graph$clustering$ids,
  .cat.feature.color = Color.Palette[1, 1:5],
  .panel.size = 2,
  .midpoint = NULL,
  .plot.title = "",
  .color.label = "",
  .legend.position = "right",
  .point.size = 0.1,
  .seed = 123,
  .hover.stats = "none",
  .hover.n.features = 10
)
}
\arguments{
\item{.lm.obj}{A list object initialized with setup.lm.obj and processed with get.landmarks and get.graph.}

\item{.feature}{The .feature to plot. It can be a numeric vector or a factor of length equal to the number of landmarks (nrow(.lm.obj$lm)). If numeric, the plot will be colored in relation to the median value of the .feature, or by .midpoint. If a factor, the plot will be colored by the levels of the factor.}

\item{.cat.feature.color}{The color palette to use in case .feature is categorical. Defaults to  colors (row 1, cols 1:5).}

\item{.panel.size}{The size of the panel.}

\item{.midpoint}{The midpoint for the color scale. Only used if .feature is numeric. Defaults to the median of the .feature.}

\item{.plot.title}{The title of the plot.}

\item{.color.label}{The label of the color scale.}

\item{.legend.position}{The position of the legend. Defaults to "right" and can be set to "left", "top", "bottom" or "none".}

\item{.point.size}{The size of the points in the plot.}

\item{.seed}{The .seed for the random number generator.}

\item{.hover.stats}{The statistics to show in the hover. Defaults to "none" and can be set to "marker" to show top features associated with each neighborhood compared to all others.}

\item{.hover.n.features}{The number of top features to show in the hover in each fold change direction. Defaults to 10 (i.e., 20 total features, 10 in each direction).}
}
\description{
This function plots the dimensionality reduction of the landmarks (UMAP).
}
\note{
Interactive hover features (when .hover.stats != "none") require the ggiraph package. If ggiraph is not available, a static plot will be returned with a warning.
}
