% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot}
\usage{
scatterPlot(
  .x.feature,
  .y.feature,
  .color.feature = NULL,
  .x.label = "",
  .y.label = "",
  .color.label = "",
  .cat.feature.color = Color.Palette[1, 1:5],
  .panel.size = c(2, 2),
  .midpoint = NULL,
  .plot.title = "",
  .legend.position = "right",
  .point.size = 0.1,
  .seed = 123
)
}
\arguments{
\item{.x.feature}{The x-axis feature. Defaults to \code{.lm.obj$scaled.lm[,"CD3"]}.}

\item{.y.feature}{The y-axis feature. Defaults to \code{.lm.obj$pca$embed[,"PC1"]}.}

\item{.color.feature}{The feature to color the plot by. Defaults to \code{.lm.obj$graph$clustering$ids}.}

\item{.x.label}{The x-axis label. Defaults to "x".}

\item{.y.label}{The y-axis label. Defaults to "y".}

\item{.color.label}{The color label. Defaults to "cluster".}

\item{.cat.feature.color}{The color palette to use in case .feature is categorical. Defaults to \code{Color.Palette[1,1:5]}.}

\item{.panel.size}{The size of the panel.}

\item{.midpoint}{The midpoint of the color gradient. Defaults to the median of .color.feature. Only used if .color.feature is numeric.}

\item{.plot.title}{The title of the plot.}

\item{.legend.position}{The position of the legend. Defaults to "right" and can be set to "left", "top", "bottom" or "none".}

\item{.point.size}{The size of the points.}

\item{.seed}{The seed for the random number generator.}
}
\value{
A ggplot2 object.
}
\description{
This function plots a scatter plot of two features colored by a third feature.
}
