% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmarks.R
\name{setup.lm.obj}
\alias{setup.lm.obj}
\title{Setup object for landmark identification and modeling}
\usage{
setup.lm.obj(
  .cells,
  .meta,
  .markers = NULL,
  .harmony.var = NULL,
  .assay.type = "cyto",
  .verbose = TRUE,
  .seed = 123,
  .prop.landmarks = 0.1,
  .n.threads = if (is.hpc()) {
     max(RhpcBLASctl::blas_get_num_procs(),
    RhpcBLASctl::omp_get_num_procs(), RhpcBLASctl::omp_get_max_threads(), na.rm = TRUE)

    } else {
     parallel::detectCores(logical = TRUE)
 }
)
}
\arguments{
\item{.cells}{A named list containing the on-disk location of RDS files containing one expression matrix for each sample. List names correspond to sample names.}

\item{.meta}{A data frame containing the sample-wise metadata.}

\item{.markers}{A character vector specifying the markers to use for landmark identification. Default is NULL. Applies only if \code{assay.type} is \code{"cyto"} since all features are included for \code{"RNA"}.}

\item{.harmony.var}{A character vector specifying the metadata variables to use for Harmony integration. Default is NULL. If provided, these variables must be present in the metadata.}

\item{.assay.type}{A character string specifying the assay type. Default is \code{"cyto"}. Must be one of \code{"cyto"} or \code{"RNA"}.}

\item{.verbose}{A logical specifying whether to print progress messages. Default is TRUE.}

\item{.seed}{An integer specifying the .seed for reproducibility. Default is 123.}

\item{.prop.landmarks}{A numeric value between 0 and 1 specifying the proportion of cells to use as landmarks. Default is 0.1 (10\% of total cells), with a maximum of 5000 landmarks.}

\item{.n.threads}{An integer specifying the number of threads to use. Default is the max allowed by the system.}
}
\value{
A list object initialized with the input parameters.
}
\description{
This function initializes an object for landmark identification and modeling.
}
