% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinydenseR-package.R
\docType{package}
\name{tinydenseR-package}
\alias{tinydenseR}
\alias{tinydenseR-package}
\title{tinydenseR: Linking Cell-To-Cell Variation to Sample-to-Sample Variation}
\description{
tinydenseR is a landmark-based platform for single cell data analysis that identifies differentially abundant cell types and differentially expressed features, including subtle within-cluster state changes. Modeling samples as replicates, tinydenseR enhances analytic efficiency and reproducibility while preserving the richness of single cell data.
}
\author{
\strong{Maintainer}: Pedro Milanez-Almeida \email{pedro.milanez@novartis.com} (\href{https://orcid.org/0000-0001-8285-957X}{ORCID})

}
\keyword{internal}
